add custom ktlint rules commit_hash:9ea8c8e634343048f16e696961dbf6e90e03e726

+++ analytics/src/test/java/io/appmetrica/analytics/impl/component/session/BaseSessionTest.java
@@ -82,7 +82,7 @@ public abstract class BaseSessionTest extends CommonTest {
     public void testSessionInValidIfExpired() throws Exception {
         mConfig.setSessionTimeout(getSessionTimeout());
 
-        mSessionStorage.putSleepStart((long) -getSessionTimeout());
+        mSessionStorage.putSleepStart(-getSessionTimeout());
 
         Session session = getSessionFactory().load();
